% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_pkg.R
\name{init_package}
\alias{init_package}
\title{My template for packages}
\usage{
init_package(
  mail = "antoine.bichat@mines-nancy.org",
  readme = TRUE,
  news = TRUE,
  test = TRUE,
  spellcheck = TRUE,
  renv = TRUE
)
}
\arguments{
\item{mail}{character}

\item{readme}{logical}

\item{news}{logical}

\item{test}{logical}

\item{spellcheck}{logical}

\item{renv}{logical}
}
\description{
My template for packages
}
\examples{
\dontrun{
init_package(news = FALSE)
}
}
